chore: Add fake google-services plugin for Travis
Firebase needs google-services.json however there is no way to use
the google-services plugin only for mock; it's all or nothing.

Using a fake google-services.json file in the mock flavour with no
API key seems to work ok so far. This way Travis is happy as gradle
will not complain, and with Travis only running mock tests it won't
need the real json file which is in prod.