fix: Day night theme in wrong spot
I noticed that the place I thought I'd put this code was actually in
the wrong class. It was meant to be in the Application class not the
Application module. Must have been asleep at the wheel.

It works fine in the correct spot so there's no need for it to be in
the ItemsActivity now.