test: Paramatised test not working
Activity not being recreated properly after the first parameter is
run. I'm not sure if it's something I'm doing or if it's a bug, but
will leave it until the next version of Espresso. Based on I/O talk
it looks like maybe this will work better with the next release.